+++ app/src/main/java/com/udacity/study/jam/radiotastic/category/CategoryListFragment.java
@@ -119,6 +119,6 @@ public class CategoryListFragment extends Fragment {
     }
 
     public static interface Callback {
-        void onCategorySelected(int categoryID);
+        void onCategorySelected(double categoryID);
     }
 }
